实现基本业务逻辑

+++ src/main/java/cn/edu/sdu/ise/labs/service/TeamService.java
@@ -46,5 +46,5 @@ public interface TeamService {
      * @param teamCodes
      * @return
      */
-    Long deleteTeam(List<String> teamCodes);
+    int deleteTeam(List<String> teamCodes);
 }
